requirements.txt
# YOLOv8 and Computer Vision
ultralytics==8.0.200
opencv-python==4.8.1.78
torch==2.1.0
torchvision==0.16.0
pillow==10.0.1
numpy==1.24.3

# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Streamlit Frontend
streamlit==1.28.1
requests==2.31.0
pandas==2.1.3

# Additional utilities
python-dotenv==1.0.0
pathlib==1.0.1